@@ -30,26 +30,26 @@ class CountryPickerDropdown extends StatefulWidget {
   });
 
   /// Filters the available country list
-  final ItemFilter itemFilter;
+  final ItemFilter? itemFilter;
 
   /// [Comparator] to be used in sort of country list
-  final Comparator<Country> sortComparator;
+  final Comparator<Country>? sortComparator;
 
   /// List of countries that are placed on top
-  final List<Country> priorityList;
+  final List<Country>? priorityList;
 
   ///This function will be called to build the child of DropdownMenuItem
   ///If it is not provided, default one will be used which displays
   ///flag image, isoCode and phoneCode in a row.
   ///Check _buildDefaultMenuItem method for details.
-  final ItemBuilder itemBuilder;
+  final ItemBuilder? itemBuilder;
 
   ///It should be one of the ISO ALPHA-2 Code that is provided
   ///in countryList map of countries.dart file.
-  final String initialValue;
+  final String? initialValue;
 
   ///This function will be called whenever a Country item is selected.
-  final ValueChanged<Country> onValuePicked;
+  final ValueChanged<Country?>? onValuePicked;
 
   /// Boolean property to enabled/disable expanded property of DropdownButton
   final bool isExpanded;
@@ -61,34 +61,34 @@ class CountryPickerDropdown extends StatefulWidget {
   final bool isDense;
 
   /// See [underline] of [DropdownButton]
-  final Widget underline;
+  final Widget? underline;
 
   /// Selected country widget builder to display. See [selectedItemBuilder] of [DropdownButton]
-  final ItemBuilder selectedItemBuilder;
+  final ItemBuilder? selectedItemBuilder;
 
   /// See [dropdownColor] of [DropdownButton]
-  final Color dropdownColor;
+  final Color? dropdownColor;
 
   /// See [onTap] of [DropdownButton]
-  final VoidCallback onTap;
+  final VoidCallback? onTap;
 
   /// See [icon] of [DropdownButton]
-  final Widget icon;
+  final Widget? icon;
 
   /// See [iconDisabledColor] of [DropdownButton]
-  final Color iconDisabledColor;
+  final Color? iconDisabledColor;
 
   /// See [iconEnabledColor] of [DropdownButton]
-  final Color iconEnabledColor;
+  final Color? iconEnabledColor;
 
   /// See [iconSize] of [DropdownButton]
   final double iconSize;
 
   /// See [hint] of [DropdownButton]
-  final Widget hint;
+  final Widget? hint;
 
   /// See [disabledHint] of [DropdownButton]
-  final Widget disabledHint;
+  final Widget? disabledHint;
 
   /// Set first item in the country list as selected initially
   /// if initialValue is not provided
@@ -99,8 +99,8 @@ class CountryPickerDropdown extends StatefulWidget {
 }
 
 class _CountryPickerDropdownState extends State<CountryPickerDropdown> {
-  List<Country> _countries;
-  Country _selectedCountry;
+  late List<Country> _countries;
+  Country? _selectedCountry;
 
   @override
   void initState() {
@@ -112,15 +112,15 @@ class _CountryPickerDropdownState extends State<CountryPickerDropdown> {
     }
 
     if (widget.priorityList != null) {
-      widget.priorityList.forEach((Country country) =>
+      widget.priorityList!.forEach((Country country) =>
           _countries.removeWhere((Country c) => country.isoCode == c.isoCode));
-      _countries.insertAll(0, widget.priorityList);
+      _countries.insertAll(0, widget.priorityList!);
     }
 
     if (widget.initialValue != null) {
       try {
         _selectedCountry = _countries.firstWhere(
-          (country) => country.isoCode == widget.initialValue.toUpperCase(),
+          (country) => country.isoCode == widget.initialValue!.toUpperCase(),
         );
       } catch (error) {
         throw Exception(
@@ -141,7 +141,7 @@ class _CountryPickerDropdownState extends State<CountryPickerDropdown> {
         .map((country) => DropdownMenuItem<Country>(
             value: country,
             child: widget.itemBuilder != null
-                ? widget.itemBuilder(country)
+                ? widget.itemBuilder!(country)
                 : _buildDefaultMenuItem(country)))
         .toList();
 
@@ -160,7 +160,7 @@ class _CountryPickerDropdownState extends State<CountryPickerDropdown> {
       onChanged: (value) {
         setState(() {
           _selectedCountry = value;
-          widget.onValuePicked(value);
+          widget.onValuePicked!(value);
         });
       },
       items: items,
@@ -169,7 +169,7 @@ class _CountryPickerDropdownState extends State<CountryPickerDropdown> {
       selectedItemBuilder: widget.selectedItemBuilder != null
           ? (context) {
               return _countries
-                  .map((c) => widget.selectedItemBuilder(c))
+                  .map((c) => widget.selectedItemBuilder!(c))
                   .toList();
             }
           : null,
